split gl3d_plot_interact_test.js
- add new gl3d_hover_click_test.js
- mv a few scatter3d-only test to scatter3d_test.js
